updated README: structure changes, new info
Restructured the "Usage" section, and added some more information on handling the 'todo's. Also expanded the Windows file generation issue to include the examples/simpletest.py file. Lastly, added "post a new issue to the Adafruit Bundle repo" as a means of contact for adding new libraries to that bundle vs the community bundle.