Add pre-commit configuration and integration for REUSE.software.
pre-commit.com provides a way to configure pre-commit hooks for multiple
tools, including `reuse lint` (configured here), isort, black, and more.

In addition to `reuse lint`, enable basic whitespace management and YAML
checks. Note that `black` is not usable with this repository because it
cannot parse the template `.py` files.

Use the pre-commit provided GitHub action as part of the CI, for ease of
maintenance.