pre-commit: run on all branches & on ubuntu-20.04
I noticed that pre-commit only runs on the 'master' branch.  This will be incorrect if cookiecutter is ever updated to use the 'main' branch designation for primary development.

I also updated the ubuntu version from ubuntu-latest to ubuntu-20.04.  This silences a warning shown in the Actions results pages that the meaning of ubuntu-latest can & will change.