Add calls to pylint from pre-commit
The pre-commit configuration file has had the calls to pylint for both
the library code (./adafruit_*.py) and example code (./examples/*.py)
added. The two calls are independant from one-another, and are named
respectively "pylint-library" and "pylint-examples".

The benefit of two separate calls is that they are easily
distinguishable from one another, their rules are applied specifically
to the correct patterns, and the comparison behaviour can be adjusted
independantly (compare all files together as a batch versus
individually).

Additionally, a shell call was the only way to work
around the pylint issue introduced in 2.7.0 which was causing duplicate
code errors to fail the validation, which should not be an issue for the
"examples/*.py" files, since each example should stand on its own.